requirements.txt
discord.py>=2.3.2
openai>=1.14.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.30
streamlit>=1.36.0
pandas>=2.2.2
psycopg2-binary==2.9.9
